conolidate to_date and to_timestamp
